Adpated package names according to previous commit. Applied fixes as requested in #234. Refactored artifact name and version as discussed in #226.
