+++ _ext/eclipse-base/gradle.properties
@@ -1,7 +1,8 @@
-# Mayor/Minor versions correspond to the minimum Eclipse version supported/tested.
-# Patch version is incremented for backward compatible patches of this library.
-ext_version=4.7.0-SNAPSHOT
-ext_artifactId=spotless-ext-eclipse-base
+# Mayor version correspond to the required Eclipse core version.
+# Minor version changes in case the minimum Eclipse core dependencies are raised due to incompatibilities of their internal interfaces.
+# Patch version is incremented in case of bug-fixes and evolutions.
+ext_version=3.0.0-SNAPSHOT
+ext_artifactId=spotless-eclipse-base
 ext_description=Eclipse bundle controller and services for Spotless
 
 ext_org=diffplug
