+++ _ext/eclipse-base/src/main/java/com/diffplug/spotless/extra/eclipse/base/runtime/PluginRegistrar.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.gradle.spotless.eclipse.runtime;
+package com.diffplug.spotless.extra.eclipse.base.runtime;
 
 import java.io.IOException;
 import java.io.InputStream;
