+++ _ext/eclipse-base/src/main/java/com/diffplug/spotless/extra/eclipse/base/service/NoDebugging.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.gradle.spotless.eclipse.service;
+package com.diffplug.spotless.extra.eclipse.base.service;
 
 import java.io.File;
 import java.util.Map;
