+++ _ext/eclipse-base/src/test/java/com/diffplug/spotless/extra/eclipse/base/SpotlessEclipseFrameworkTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.gradle.spotless.eclipse;
+package com.diffplug.spotless.extra.eclipse.base;
 
 import org.junit.Test;
 import org.osgi.framework.BundleException;
